updated Factor to be compatible with new Jacobian api
