+++ autofit/graphical/factor_graphs/jacobians.py
@@ -19,154 +19,25 @@ from autofit.graphical.factor_graphs.factor import (
     Factor,
     DeterministicFactor,
 )
-from autofit.graphical.utils import aggregate, Axis, nested_filter, nested_update
+
+from autofit.graphical.factor_graphs.abstract import (
+    AbstractJacobian,
+    JacobianVectorProduct,
+    VectorJacobianProduct,
+)
+from autofit.graphical.utils import (
+    aggregate,
+    Axis,
+    nested_filter,
+    nested_update,
+    is_variable,
+)
 from autofit.mapper.variable import (
     Variable,
     Plate,
-    VariableLinearOperator,
     VariableData,
     FactorValue,
 )
-from autofit.mapper.variable_operator import (
-    RectVariableOperator,
-    LinearOperator,
-    VariableOperator,
-)
-
-
-def _is_variable(v, *args):
-    return isinstance(v, Variable)
-
-
-class AbstractJacobian(VariableLinearOperator):
-    """
-    Examples
-    --------
-    def linear(x, a, b):
-        z = x.dot(a) + b
-        return (z**2).sum(), z
-
-    def full(x, a, b):
-        z2, z = linear(x, a, b)
-        return z2 + z.sum()
-
-    x_, a_, b_, y_, z_ = variables("x, a, b, y, z")
-    x = np.arange(10.).reshape(5, 2)
-    a = np.arange(2.).reshape(2, 1)
-    b = np.ones(1)
-    y = np.arange(0., 10., 2).reshape(5, 1)
-    # values = {x_: x, y_: y, a_: a, b_: b}
-
-    linear_factor_jvp = FactorJVP(
-        linear, x_, a_, b_, factor_out=(FactorValue, z_))
-
-    linear_factor_vjp = FactorVJP(
-        linear, x_, a_, b_, factor_out=(FactorValue, z_))
-
-    values = {x_: x, a_: a, b_: b}
-
-    jvp_val, jvp_jac = linear_factor_jvp.func_jacobian(values)
-    vjp_val, vjp_jac = linear_factor_vjp.func_jacobian(values)
-
-
-    assert np.allclose(vjp_val, jvp_val)
-    assert (vjp_jac(vjp_val) - jvp_jac(vjp_val)).norm() == 0
-    """
-
-    def __call__(self, values):
-        return self.__rmul__(values)
-
-    def __str__(self) -> str:
-        out_var = str(
-            nested_update(self.factor_out, {v: v.name for v in self.out_variables})
-        ).replace("'", "")
-
-        in_var = ", ".join(v.name for v in self.variables)
-        cls_name = type(self).__name__
-        return f"{cls_name}({out_var} → ∂({in_var})ᵀ {out_var})"
-
-    __repr__ = __str__
-
-    def _full_repr(self) -> str:
-        out_var = str(self.factor_out)
-        in_var = str(self.variables)
-        cls_name = type(self).__name__
-        return f"{cls_name}({out_var} → ∂({in_var})ᵀ {out_var})"
-
-    def grad(self, values=None):
-        grad = VariableData({FactorValue: 1.0})
-        if values:
-            grad.update(values)
-
-        for v, g in self(grad).items():
-            grad[v] = grad.get(v, 0) + g
-
-        return grad
-
-
-class JacobianVectorProduct(AbstractJacobian, RectVariableOperator):
-    __init__ = RectVariableOperator.__init__
-
-    @property
-    def variables(self):
-        return self.left_variables
-
-    @property
-    def out_variables(self):
-        return self.right_variables
-
-    @property
-    def factor_out(self):
-        return tuple(self.out_variables)
-
-
-class VectorJacobianProduct(AbstractJacobian):
-    def __init__(
-        self, factor_out, vjp: Callable, *variables: Variable, out_shapes=None
-    ):
-        self.factor_out = factor_out
-        self.vjp = vjp
-        self._variables = variables
-        self.out_shapes = out_shapes
-
-    @property
-    def variables(self):
-        return self._variables
-
-    @cached_property
-    def out_variables(self):
-        return set(v[0] for v in nested_filter(_is_variable, self.factor_out))
-
-    def _get_cotangent(self, values):
-        if isinstance(values, FactorValue):
-            values = values.to_dict()
-
-        if isinstance(values, dict):
-            if self.out_shapes:
-                for v in self.out_shapes.keys() - values.keys():
-                    values[v] = np.zeros(self.out_shapes[v])
-            out = nested_update(self.factor_out, values)
-            return out
-
-        if isinstance(values, int):
-            values = float(values)
-
-        return values
-
-    def __call__(self, values: Union[VariableData, FactorValue]) -> VariableData:
-        v = self._get_cotangent(values)
-        grads = self.vjp(v)
-        return VariableData(zip(self.variables, grads))
-
-    __rmul__ = __call__
-
-    def _not_implemented(self, *args):
-        raise NotImplementedError()
-
-    __rtruediv__ = _not_implemented
-    ldiv = _not_implemented
-    __mul__ = _not_implemented
-    update = _not_implemented
 
 
 class FactorJac(Factor):
@@ -247,7 +118,7 @@ class FactorJac(Factor):
         name = name or factor.__name__
         AbstractFactor.__init__(self, **kwargs, name=name, plates=plates)
 
-        det_variables = set(v[0] for v in nested_filter(_is_variable, factor_out))
+        det_variables = set(v[0] for v in nested_filter(is_variable, factor_out))
         det_variables.discard(FactorValue)
         self._deterministic_variables = det_variables
 
@@ -302,9 +173,7 @@ class FactorJac(Factor):
         where the values of the deterministic values are stored in a dict
         attribute `FactorValue.deterministic_values`
         """
-        det_values = VariableData(
-            nested_filter(_is_variable, self.factor_out, raw_fval)
-        )
+        det_values = VariableData(nested_filter(is_variable, self.factor_out, raw_fval))
         fval = det_values.pop(FactorValue, 0.0)
         return FactorValue(fval, det_values)
 
@@ -363,19 +232,12 @@ class FactorJac(Factor):
         factor._factor(*args), jax.jacobian(factor._factor, range(len(args)))(*args)
         """
         eps = eps or self.eps
-        # args = tuple(np.array(values[v]) for v in self.args)
         args = tuple(np.array(value) for value in args)
 
         raw_fval0 = self._factor(*args)
         fval0 = self._factor_value(raw_fval0).to_dict()
-        # in_shapes = {v: np.shape(a) for v, a in values.items()}
-        # out_shapes = {v: np.shape(a) for v, a in fval0.items()}
 
         jac = {
-            # v0: {
-            #     v1: np.empty_like(fval0[v0], shape=shape0 + shape1)
-            #     for v1, shape1 in in_shapes.items()
-            # }
             v0: tuple(
                 np.empty_like(val, shape=np.shape(val) + np.shape(value))
                 for value in args
@@ -401,7 +263,7 @@ class FactorJac(Factor):
 
     def _unpack_jacobian_out(self, raw_jac: Any) -> Dict[Variable, VariableData]:
         jac = {}
-        for v0, vjac in nested_filter(_is_variable, self.factor_out, raw_jac):
+        for v0, vjac in nested_filter(is_variable, self.factor_out, raw_jac):
             jac[v0] = VariableData()
             for v1, j in zip(self.args, vjac):
                 jac[v0][v1] = j
